requirements_minimal.txt
# Minimal requirements for PythonAnywhere free account
# Optimized to reduce disk space usage

# Core Flask
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0

# Google AI (essential only)
google-generativeai==0.8.5
google-auth==2.40.3
google-auth-httplib2==0.2.0

# HTTP requests
requests==2.32.5
urllib3==2.5.0

# Environment variables
python-dotenv==1.0.1

# CORS support
flask-cors==4.0.1

# Production server
gunicorn==23.0.0
